split up common reimplementation check #415
